Change hostmaster to dns admin based on this email? (Micronet)
From: Isaac Orr <iorr@berkeley.edu>
To:   ucb-net-announce lists.berkeley.edu <ucb-net-announce@lists.berkeley.edu>, group Micronet-UCB user microcomputer support <micronet-list@lists.berkeley.edu>, OneIT@calgroups.berkeley.edu
Date: Friday, February 5 2021 at 8:02 AM PST

Subject: [Micronet] ANNOUNCEMENT: Renaming of "Hostmaster" Role to "DNS Administrator"

Based on careful consideration of the OneIT goal of creating a diverse and inclusive community of IT professionals, the NOS team has evaluated the naming of the well known “hostmaster” role and service. We do not believe the current name for this role and service is consistent with our values as we understand them today.

Considering the identity of the current member of our community who primarily performs this role, the current name is especially inaccurate. Indeed, many of us refer to that person as “hostmistress” in acknowledgement of this fact.

This term of course still carries connotations which are not acceptable if we are serious about making our area of practice more diverse and inclusive.

The “hostmaster” name is one which historically is defined in actual standards for internet operation and deeply embedded in various systems. As such we do not expect that we will be able to eliminate all references to it in the short or medium term. We do however intend to update our documentation and deprecate the use of “hostmaster” or “hostmistress” wherever it is in common usage.

Our intention is to replace this wording with “DNS Administrator” which we believe is a more inclusive, accurate representation of the role which will be well understood by our community.

The hostmaster@berkeley.edu email address, commonly used for requesting DNS and IP addressing services, will be retired effective immediately. It is replaced by dns@berkeley.edu which is already fully operational.

The old email address will continue to function, but we would like to request your support in ensuring that any of your own processes, documentation and work practices are updated to use the new address and role title.

--
Isaac Simon Orr
Manager, Network Operations and Services
IST Telecommunications, UC Berkeley
P: +1 510 643 9837 C: +1 510 517 9408 E: iorr@berkeley.edu

--
-------------------------------------------------------------------------
The following was automatically added to this message by the list server:

To learn more about Micronet, please visit the Micronet Web site: http://micronet.berkeley.edu [1]

Messages you send to this mailing list are public and world-viewable, and the list's archives can be browsed and searched on the Internet. This means these messages can be viewed by (among others) your bosses, prospective employers, and people who have known you in the past.

ANNOUNCEMENTS: To send announcements to the Micronet list, please use the micronet-announce@lists.berkeley.edu list.
---
You received this message because you are subscribed to the Google Groups "Micronet" group.
To unsubscribe from this group and stop receiving emails from it, send an email to micronet-list+unsubscribe@lists.berkeley.edu.
To view this discussion on the web visit https://groups.google.com/a/lists.berkeley.edu/d/msgid/micronet-list/CAEDbbdChzEFFVhkE1viBxHyAhz%2BwiN3ENktKh5S899p%3DKukjBg%40mail.gmail.com [2].

--
[1] http://micronet.berkeley.edu/
[2] https://groups.google.com/a/lists.berkeley.edu/d/msgid/micronet-list/CAEDbbdChzEFFVhkE1viBxHyAhz%2BwiN3ENktKh5S899p%3DKukjBg%40mail.gmail.com?utm_medium=email&utm_source=footer